fixes for draft notification and changed api call for notifications
